+++ crates/stackable-operator/src/commons/mod.rs
@@ -13,3 +13,4 @@ pub mod resources;
 pub mod s3;
 pub mod secret;
 pub mod secret_class;
+pub mod tls_verification;